+++ app/exceptions/matched_concept_unpublished.rb
@@ -1,0 +1,7 @@
+class MatchedConceptUnpublished < ArgumentError
+  attr_reader :original
+  def initialize(msg, original = $!)
+    super(msg)
+    @original = original;
+  end
+end