+++ app/models/dataset/adaptors/iqvoc/alphabetical_search_adaptor.rb
@@ -13,8 +13,7 @@ class Dataset::Adaptors::Iqvoc::AlphabeticalSearchAdaptor < Dataset::Adaptors::I
         fetch_results(more[:href], {})
       end
     rescue Faraday::Error::ConnectionFailed,
-      Faraday::Error::ResourceNotFound,
-      Faraday::Error::TimeoutError => e
+      Faraday::Error::ResourceNotFound => e
         Rails.logger.warn("HTTP error while querying remote source #{url}: #{e.message}")
         return nil
     end
@@ -29,7 +28,8 @@ class Dataset::Adaptors::Iqvoc::AlphabeticalSearchAdaptor < Dataset::Adaptors::I
 
       options = {
         definition: element.at_css('.concept-item-definition').try(:content),
-        definition_language: element.at_css('.concept-item-definition').try(:[], :lang)
+        definition_language: element.at_css('.concept-item-definition').try(:[], :lang),
+        additional_info: element.at_css('.additional_info').try(:content)
       }
 
       result = AlphabeticalSearchResultRemote.new(url, path, label, options)